Reorganized tree in preparation for beta release. Split simple (example) jack host and more useful one (with midi). Working Jack/LV2 midi in lv2_jack_host. Added lv2_list.
git-svn-id: http://svn.drobilla.net/lad/slv2@279 a436a847-0d15-0410-975c-d299462d15a1

+++ hosts/lv2-midifunctions.h
@@ -148,7 +148,7 @@ static int lv2midi_put_event(LV2_MIDIState* state,
   state->midi->size += sizeof(double);
   *(size_t*)(state->midi->data + state->midi->size) = size;
   state->midi->size += sizeof(size_t);
-  memcpy(state->midi->data + state->midi->size, data, size);
+  memcpy(state->midi->data + state->midi->size, data, (size_t)size);
   state->midi->size += size;
   
   ++state->midi->event_count;
